Add simple example of writing RawAudio messages to an mcap. (#431)
### Changelog
Add a RawAudio example that writes out a mono sinewave to an mcap.

### Docs

None

### Description
Adds a simple python example of writing RawAudio messages to an mcap
file. This just writes a single channel sine wave to keep things as
simple as possible. We could consider adding an example that writes a
stereo file to help people understand how to write interleaved samples
if we think that would be useful but it would increase the complexity
slightly.

---------

Co-authored-by: Miles Egan <miles@foxglove.dev>