Don't use default values from schemas for C++ (#436)
This removes the non-zero default values for C++ schemas to match Rust
and Python.

The reasoning:

There weren't many fields with a non-zero default value, and while w=1.0
makes sense for quarternion, the values make less sense for vectors and
colors.

Rust has default of 0 as defined by prost protobuf library, and we can't
change that, so if we use non-zero defaults elsewhere it won't be
consistent across the sdk.

Having everything default to 0, but a couple exceptions could be
surprising to developers.

We didn't document the defaults:
https://docs.foxglove.dev/docs/visualization/message-schemas/quaternion

Slack discussion:
https://foxglove.slack.com/archives/C07RB3SSJE5/p1747846188213449