Add a c++ example using the typed channels and foxglove (#438)
Add an example of using the C++ typed channels, similar to the box and
frame transform example in Rust.

The code works and shows more or less how to use the channels, but
opening the mcap in the foxglove app is very much underwhelming. I think
that cold be improved on, and I lack the domain specific knowledge on
how to do that.