codegen: Remove "Primitive" suffix filters (#442)
When we wrote `foxglove_proto_gen` back in the day, we excluded certain
kinds of codegen for schema types whose names end in "Primitive". The
choice was rather arbitrary, and it's been questioned a few times in
various contexts.

I finally bumped up against it while trying to plumb through a
`get_schema()` staticmethod for schema types in python, so I hereby
declare that it's time to drop the filter. If you want to create an
`ArrowPrimitive` channel, go for it. It's surely no stranger than a
`Color` channel.