python: Add Message.get_schema() and various channel getters (#441)
This change add a `get_schema()` staticmethod for foxglove message
types, and plumbs various getter methods for channels:
- message_encoding
- schema
- metadata
- has_sinks

Since `message_encoding` was previously an instance attribute on
`Channel`, I decided to convert them to a instance property method for
compatibility reasons, even though it's not consistent with the instance
methods we use for topic(), id(), etc.

This change also derives the magic `eq` methods for the `Schema` class. 

It's a big change, but it's mostly codegen.