C++ fix assertion crash copying zero-length vectors via the arena (#461)
### Changelog

Fix assertion failure in C++ when serializing empty vectors

### Description

arena.hpp:73: T *foxglove::Arena::alloc(size_t) [T =
foxglove_scene_entity_deletion]: Assertion `elements > 0' failed.

The problem was that arena alloc expects a non-zero allocation, but
arena map was calling it without checking the number of elements for 0
first. This checks and returns a nullptr for empty arrays. This means
they're passed to C as a nullptr with length 0, which fits with how we
represent empty arrays and strings.

While I was at it, I noticed that forward declaring context and errors
in mcap.hpp meant that you had to include both of those headers whenever
you import mcap.hpp, I changed this to match how we did channel.hpp and
server.hpp, which makes it a little more friendly to use.