Log struct with jsonschema & msgpack (#464)
This adds a C++ example which illustrates logging of custom data without
manual serialization.

The example sets up two channels; one with a json schema derived from
the struct, and one using [msgpack](https://msgpack.org/). The latter is
schemaless, but provides a binary option. The Foxglove app handles this
in the same way as schemaless JSON. A handful of messages are written to
an MCAP file. Both techniques rely on https://json.nlohmann.me/, which
is also used by the ws-services example.

This provides a couple of options for users who want to put together
logging of custom schemas quickly, without caring about the
serialization format. In python, we accomplish this by logging native
types (dict, etc); in rust, we provide derive macros for `Encode`, or
illustrate usage with `schemars::JsonSchema`.